+++ Source/equipment/battleaxe-thirsting.js
@@ -1,14 +1,14 @@
 const EquipmentTemplate = require('../../Classes/EquipmentTemplate.js');
 const { addModifier, dealDamage, gainHealth, getFullName } = require('../combatantDAO.js');
 
-module.exports = new EquipmentTemplate("Thirsting Battleaxe", "Strike a foe for @{damage} (+@{bonusDamage} if you have 0 block) @{element} damage, gain @{healing} hp on kill", "Damage x@{critBonus}", "Fire", effect, ["Prideful Battleaxe", "Thick Battleaxe"])
+module.exports = new EquipmentTemplate("Thirsting Battleaxe", "Strike a foe for @{damage} (+@{bonus} if you have 0 block) @{element} damage, gain @{healing} hp on kill", "Damage x@{critBonus}", "Fire", effect, ["Prideful Battleaxe", "Thick Battleaxe"])
 	.setCategory("Weapon")
 	.setTargetingTags({ target: "single", team: "enemy" })
 	.setModifiers([{ name: "Stagger", stacks: 1 }])
 	.setCost(350)
 	.setUses(10)
 	.setDamage(50)
-	.setBonusDamage(100)
+	.setBonus(100) // damage
 	.setHealing(60);
 
 function effect([target], user, isCrit, adventure) {