+++ Source/equipment/censer-thick.js
@@ -2,12 +2,12 @@ const EquipmentTemplate = require('../../Classes/EquipmentTemplate.js');
 const { dealDamage, addModifier, getFullName } = require('../combatantDAO.js');
 const { isDebuff } = require('../Modifiers/_modifierDictionary.js');
 
-module.exports = new EquipmentTemplate("Thick Censer", "Burn a foe for @{damage} (+@{bonusDamage} if target has any debuffs) @{element} damage", "Also apply @{mod1Stacks} @{mod1}", "Fire", effect, ["Tormenting Censor"])
+module.exports = new EquipmentTemplate("Thick Censer", "Burn a foe for @{damage} (+@{bonus} if target has any debuffs) @{element} damage", "Also apply @{mod1Stacks} @{mod1}", "Fire", effect, ["Tormenting Censor"])
 	.setCategory("Trinket")
 	.setTargetingTags({ target: "single", team: "enemy" })
 	.setModifiers([{ name: "Stagger", stacks: 1 }, { name: "Slow", stacks: 2 }])
 	.setDamage(50)
-	.setBonusDamage(75)
+	.setBonus(75) // damage
 	.setCost(350)
 	.setUses(20);