+++ Source/equipment/daggers-wicked.js
@@ -1,7 +1,7 @@
 const EquipmentTemplate = require('../../Classes/EquipmentTemplate.js');
 const { dealDamage, addModifier, getFullName } = require("../combatantDAO.js");
 
-module.exports = new EquipmentTemplate("Wicked Daggers", "Strike a foe for @{damage} (+@{bonusDamage} if foe has 0 block) @{element} damage", "Damage x@{critBonus}", "Wind", effect, ["Sharpened Daggers", "Sweeping Daggers"])
+module.exports = new EquipmentTemplate("Wicked Daggers", "Strike a foe for @{damage} (+@{bonus} if foe has 0 block) @{element} damage", "Damage x@{critBonus}", "Wind", effect, ["Sharpened Daggers", "Sweeping Daggers"])
 	.setCategory("Weapon")
 	.setTargetingTags({ target: "single", team: "enemy" })
 	.setModifiers([{ name: "Stagger", stacks: 1 }])
@@ -9,7 +9,7 @@ module.exports = new EquipmentTemplate("Wicked Daggers", "Strike a foe for @{dam
 	.setUses(10)
 	.setCritBonus(3)
 	.setDamage(75)
-	.setBonusDamage(50);
+	.setBonus(50); // damage
 
 function effect([target], user, isCrit, adventure) {
 	if (target.hp < 1) {