+++ Source/equipment/inspiration-base.js
@@ -1,22 +1,20 @@
 const EquipmentTemplate = require('../../Classes/EquipmentTemplate.js');
 const { addModifier, removeModifier } = require('../combatantDAO.js');
 
-module.exports = new EquipmentTemplate("Inspiration", "Apply @{mod1Stacks} @{mod1} to an ally", "@{mod2} x@{critBonus}", "Wind", effect, ["Soothing Inspiration", "Sweeping Inspiration"])
+module.exports = new EquipmentTemplate("Inspiration", "Apply @{mod1Stacks} @{mod1} to an ally", "@{mod1} +@{bonus}", "Wind", effect, ["Soothing Inspiration", "Sweeping Inspiration"])
 	.setCategory("Spell")
 	.setTargetingTags({ target: "single", team: "delver" })
-	.setModifiers([{ name: "Stagger", stacks: 1 }, { name: "Power Up", stacks: 25 }, { name: "Power Up", stacks: 50 }])
+	.setModifiers([{ name: "Stagger", stacks: 1 }, { name: "Power Up", stacks: 25 }])
+	.setBonus(25) // Power Up stacks
 	.setCost(200)
 	.setUses(5);
 
 function effect([target], user, isCrit, adventure) {
-	let { element, modifiers: [elementStagger, powerUp, critPowerUp] } = module.exports;
+	let { element, modifiers: [elementStagger, powerUp], bonus } = module.exports;
+	const pendingPowerUp = { ...powerUp, stacks: powerUp.stacks + (isCrit ? bonus : 0) };
 	if (user.element === element) {
 		removeModifier(target, elementStagger);
 	}
-	if (isCrit) {
-		addModifier(target, critPowerUp);
-	} else {
-		addModifier(target, powerUp);
-	}
+	addModifier(target, pendingPowerUp);
 	return "";
 }