+++ Source/equipment/potionkit-guarding.js
@@ -12,14 +12,14 @@ module.exports = new EquipmentTemplate("Guarding Potion Kit", "Gain @{block} blo
 	.setBlock(75);
 
 function effect(targets, user, isCrit, adventure) {
-	let { element, modifiers: [elementStagger], block } = module.exports;
+	let { element, modifiers: [elementStagger], block, critBonus } = module.exports;
 	if (user.element === element) {
 		removeModifier(user, elementStagger);
 	}
 	const randomPotion = rollConsumable(adventure, "Potion");
 	addBlock(user, block);
 	if (isCrit) {
-		adventure.addResource(new Resource(randomPotion, "consumable", 2, "loot", 0));
+		adventure.addResource(new Resource(randomPotion, "consumable", critBonus, "loot", 0));
 		return `${getFullName(user, adventure.room.enemyTitles)} hunkers down and sets a double-batch of ${randomPotion} simmering.`;
 	} else {
 		adventure.addResource(new Resource(randomPotion, "consumable", 1, "loot", 0));