+++ Source/equipment/spear-reactive.js
@@ -1,14 +1,14 @@
 const EquipmentTemplate = require('../../Classes/EquipmentTemplate.js');
 const { dealDamage, addModifier, calculateTotalSpeed, getFullName } = require("../combatantDAO.js");
 
-module.exports = new EquipmentTemplate("Reactive Spear", "Strike a foe for @{damage} (+@{bonusDamage} if foe went first) @{element} damage", "Also inflict @{mod1Stacks} @{mod1}", "Fire", effect, ["Lethal Spear", "Sweeping Spear"])
+module.exports = new EquipmentTemplate("Reactive Spear", "Strike a foe for @{damage} (+@{bonus} if foe went first) @{element} damage", "Also inflict @{mod1Stacks} @{mod1}", "Fire", effect, ["Lethal Spear", "Sweeping Spear"])
 	.setCategory("Weapon")
 	.setTargetingTags({ target: "single", team: "enemy" })
 	.setModifiers([{ name: "Stagger", stacks: 1 }, { name: "Stagger", stacks: 1 }])
 	.setCost(350)
 	.setUses(10)
 	.setDamage(100)
-	.setBonusDamage(75);
+	.setBonus(75); // damage
 
 function effect([target], user, isCrit, adventure) {
 	if (target.hp < 1) {