+++ Source/equipment/warhammer-piercing.js
@@ -1,14 +1,14 @@
 const EquipmentTemplate = require('../../Classes/EquipmentTemplate.js');
 const { dealDamage, addModifier, getFullName } = require('../combatantDAO.js');
 
-module.exports = new EquipmentTemplate("Piercing Warhammer", "Strike a foe for @{damage} (+@{bonusDamage} if foe is already stunned) unblockable @{element} damage", "Damage x@{critBonus}", "Earth", effect, [])
+module.exports = new EquipmentTemplate("Piercing Warhammer", "Strike a foe for @{damage} (+@{bonus} if foe is already stunned) unblockable @{element} damage", "Damage x@{critBonus}", "Earth", effect, ["Slowing Warhammer"])
 	.setCategory("Weapon")
 	.setTargetingTags({ target: "single", team: "enemy" })
 	.setModifiers([{ name: "Stagger", stacks: 1 }])
 	.setCost(350)
 	.setUses(10)
-	.setDamage(100)
-	.setBonusDamage(75);
+	.setDamage(75)
+	.setBonus(75); // damage
 
 function effect([target], user, isCrit, adventure) {
 	if (target.hp < 1) {